% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjugate_update_gamma.R
\name{conjugate_update_gamma_precision_matrix}
\alias{conjugate_update_gamma_precision_matrix}
\title{Conjugate update for the parameters of ind. Gamma distributions
with ind. mvn child nodes that have a mvn prior mean; prior precision
of child nodes}
\usage{
conjugate_update_gamma_precision_matrix(data, pars, priors)
}
\arguments{
\item{data}{List with \verb{S, H}}

\item{pars}{List with \verb{phi, theta}}

\item{priors}{List with \verb{a, b}}
}
\description{
Conjugate update for the parameters of ind. Gamma distributions
with ind. mvn child nodes that have a mvn prior mean; prior precision
of child nodes
}
