% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjugate_update_mvn.R
\name{conjugate_update_mvn_hierarchical_mean}
\alias{conjugate_update_mvn_hierarchical_mean}
\title{Conjugate update for the parameters of multivariate normal
distribution for the hierarchical mean of
multivariate normal child nodes}
\usage{
conjugate_update_mvn_hierarchical_mean(data, pars)
}
\arguments{
\item{pars}{List with \code{phi} and \code{theta}}
}
\description{
Conjugate update for the parameters of multivariate normal
distribution for the hierarchical mean of
multivariate normal child nodes
}
