% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentials_reh.R
\name{d_mvn_cov_reh}
\alias{d_mvn_cov_reh}
\title{The contribution of the likelihood to the
differential w.r.t. vec of the mvn covariance,
for Poisson-distributed data}
\usage{
d_mvn_cov_reh(data, pars)
}
\arguments{
\item{data}{List of observed or known variables, including 'C'.}

\item{pars}{List of parameters, including \code{phi$mu}, \code{phi$Sigma},}
}
\description{
The contribution of the likelihood to the
differential w.r.t. vec of the mvn covariance,
for Poisson-distributed data
}
