% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_differentials.R
\name{d_mvn_mean}
\alias{d_mvn_mean}
\title{Transpose of the differential w.r.t. mvn mean}
\usage{
d_mvn_mean(data, pars, d_likelihood)
}
\arguments{
\item{data}{List of observed or known variables, including 'C', 'y'.
If \code{d_likelihood()} requires only a subset of \code{data} param
given to \code{ncvi()}, \code{data} should be only this subset.}

\item{pars}{List of parameters, including \code{theta$Tau}, \code{theta$M},
\code{phi$mu}, \code{phi$Sigma}.
If \code{d_likelihood()} requires only a subset of the \code{pars} param
given to \code{ncvi()}, \code{pars} should be only this subset.}

\item{d_likelihood}{Function to provide the contribution of the
data likelihood to the differential. Should take params \code{data}
and \code{pars} as given to this function.}
}
\description{
Transpose of the differential w.r.t. mvn mean
}
