% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_differentials.R
\name{d_mvn_mean_poisson}
\alias{d_mvn_mean_poisson}
\title{The contribution of the likelihood to the
differential w.r.t. mvn mean,
for Poisson-distributed data}
\usage{
d_mvn_mean_poisson(data, pars)
}
\arguments{
\item{data}{List of observed or known variables, including 'C'.}

\item{pars}{List of parameters, including \code{phi$mu}, \code{phi$Sigma},}
}
\description{
The contribution of the likelihood to the
differential w.r.t. mvn mean,
for Poisson-distributed data
}
