% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex4_legacy.R
\name{differential_EB}
\alias{differential_EB}
\title{Legacy differential for mvn mean}
\usage{
differential_EB(data, pars)
}
\arguments{
\item{data}{List with \verb{etc$X, y, etc$t}}

\item{pars}{List with \verb{phi$EB, phi$SigmaB}}
}
\description{
Legacy differential for mvn mean
}
