% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion_est.R
\name{dispersion_est}
\alias{dispersion_est}
\title{A function to calculate tagwise dispersion based on empirical
function from real data}
\usage{
dispersion_est(beta0)
}
\description{
A function to calculate tagwise dispersion based on empirical
function from real data
}
