% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbo_hierarchical.R
\name{elbo_extra}
\alias{elbo_extra}
\title{Contribution to ELBO of terms not included in group-wise ELBO}
\usage{
elbo_extra(data, pars)
}
\arguments{
\item{data}{List of \verb{y, C, G}}

\item{pars}{List of \verb{phi, theta}}
}
\description{
Contribution to ELBO of terms not included in group-wise ELBO
}
