% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex4functions.R
\name{elbo_extra_ex4}
\alias{elbo_extra_ex4}
\title{Extra ELBO component outside group contribution}
\usage{
elbo_extra_ex4(data, pars)
}
\arguments{
\item{data}{The entire \code{data} arg}

\item{pars}{The entire \code{pars} arg}
}
\description{
Extra ELBO component outside group contribution
}
