% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex4_legacy.R
\name{elbo_extra_legacy}
\alias{elbo_extra_legacy}
\title{Extra ELBO component outside group contribution}
\usage{
elbo_extra_legacy(data, pars)
}
\arguments{
\item{data}{The entire \code{data} arg}

\item{pars}{The entire \code{pars} arg}
}
\description{
Extra ELBO component outside group contribution
}
