% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbo_hierarchical.R
\name{elbo_hierarchical}
\alias{elbo_hierarchical}
\title{Complete ELBO function}
\usage{
elbo_hierarchical(data, pars)
}
\arguments{
\item{data}{List of \verb{y, C, G}}

\item{pars}{List of \verb{phi, theta}}
}
\description{
Complete ELBO function
}
