% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbo_hierarchical.R
\name{elbo_i}
\alias{elbo_i}
\title{Contribution to EBLO from a single group}
\usage{
elbo_i(data, pars)
}
\arguments{
\item{data}{List of \verb{y, C, G}}

\item{pars}{List of \verb{phi, theta}}
}
\description{
Contribution to EBLO from a single group
}
