% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex4_legacy.R
\name{elbo_i_legacy}
\alias{elbo_i_legacy}
\title{ELBO contribution for a single group}
\usage{
elbo_i_legacy(data, pars)
}
\arguments{
\item{data}{List with \verb{etc$X, y, etc$t}}

\item{pars}{List with \verb{phi$EB, phi$SigmaB}}
}
\description{
ELBO contribution for a single group
}
