% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex4_legacy.R
\name{elbo_legacy}
\alias{elbo_legacy}
\title{Entire ELBO function for ex4}
\usage{
elbo_legacy(data, pars)
}
\arguments{
\item{data}{The entire \code{data} arg}

\item{pars}{The entire \code{pars} arg}
}
\description{
Entire ELBO function for ex4
}
