% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_babel.R
\name{fit_babel}
\alias{fit_babel}
\title{Function to fit the \code{babel} model to a RiboSeq experiment}
\usage{
fit_babel(data_babel, nreps, mc.cores = 1)
}
\arguments{
\item{data_babel}{Output of \code{prep_babel()}}

\item{nreps}{Integer, number of mc reps}

\item{mc.cores}{Integer, to set \code{options(mc.cores = ...)}}
}
\value{
A list, the output of \code{babel::babel()}
}
\description{
Function to fit the \code{babel} model to a RiboSeq experiment
}
