% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_deseq2.R
\name{fit_deseq}
\alias{fit_deseq}
\title{Function to fit the \code{DESeq2} model to a RiboSeq experiment
and return results for the interaction effect}
\usage{
fit_deseq(data_deseq)
}
\arguments{
\item{`data_deseq`}{A list as produced by \code{prep_deseq()}}
}
\value{
A list with \code{dds} (a \code{DESeq} data object with
attached results) and \code{res} from \code{results()} with the interaction effect
parameter
}
\description{
Function to fit the \code{DESeq2} model to a RiboSeq experiment
and return results for the interaction effect
}
