% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ncvi.R
\name{fit_ncvi}
\alias{fit_ncvi}
\title{Non-conjugate variational inference}
\usage{
fit_ncvi(
  data,
  init,
  differentials,
  update_phi,
  update_theta,
  elbo,
  options = list(max_iter = 100, elbo_delta = 0.001, verbose = T)
)
}
\arguments{
\item{data}{List of observed or known variables used by the
differentials or update functions.}

\item{init}{List of initial values for unknown variables
updated by the algorithm. Should have 'phi' and 'theta'.}

\item{differentials}{List of functions for differentials used in
non-conjugate updates of 'phi' parameters.}

\item{update_phi}{Function to carry out non-conjugate updates of
'phi' parameter. Should take args 'data', 'pars', and 'differentials'.}

\item{update_theta}{Function to carry out conjugate updates of 'theta'
parameter. Should take args 'data' and 'pars'.}

\item{elbo}{Function to calculate the ELBO for the model.}

\item{options}{List of options.
Should include positive real 'elbo_delta', the threshold
change in ELBO to terminate the algorithm.}
}
\description{
General form of wrapper function for a
variational inference algorithm that may include
both conjugate and non-conjugate conditional distributions.
}
