% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_xtail.R
\name{fit_xtail}
\alias{fit_xtail}
\title{Function to fit \code{xtail} model to paired RiboSeq and RNA count data}
\usage{
fit_xtail(data_xtail, bins, threads = 2, normalize = T, minMeanCount = 1)
}
\arguments{
\item{data_xtail}{list with \code{rpf} and \code{mrna} data frames and
character vector \code{condition} of treatment conditions}

\item{bins}{\code{bins} argument for \code{xtail()}}

\item{threads}{\code{threads} argument for \code{xtail()}}

\item{normalize}{\code{normalize} argument for \code{xtail()}}
}
\value{
\code{xtail} fit object
}
\description{
Function to fit \code{xtail} model to paired RiboSeq and RNA count data
}
