% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_init.R
\name{generate_inits_REH}
\alias{generate_inits_REH}
\title{Generate a set of inits for hierarchical model, with random effects}
\usage{
generate_inits_REH(
  G,
  N,
  P,
  U,
  precision_mu0,
  precision_beta,
  precision_u,
  mu_adjust = 0
)
}
\description{
Generate a set of inits for hierarchical model, with random effects
}
