% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ci.R
\name{make_ci_list}
\alias{make_ci_list}
\title{make list of lists with means, lower bound,
and upper bound for 1-alpha confidence intervals
for params in \code{pars$phi}}
\usage{
make_ci_list(fit, alpha = 0.05)
}
\arguments{
\item{fit}{A fit produced by \code{fit_ncvi()}}

\item{alpha}{Confidence level}
}
\description{
make list of lists with means, lower bound,
and upper bound for 1-alpha confidence intervals
for params in \code{pars$phi}
}
