% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_inference.R
\name{make_test_list}
\alias{make_test_list}
\title{Perform a hypothesis test based on posterior}
\usage{
make_test_list(fit, cutoff)
}
\arguments{
\item{fit}{Fit from \code{fit_ncvi()}}

\item{cutoff}{Threshold value}
}
\description{
Perform a hypothesis test based on posterior
}
