% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_xtail.R
\name{make_xtail_data}
\alias{make_xtail_data}
\title{Function to take the simulated data set and format it
as required for \code{xtail}}
\usage{
make_xtail_data(sim)
}
\arguments{
\item{sim}{output of \code{sim_data_mixture()}}
}
\value{
list with RPF and mRNA data frames and
vector of treatment conditions
}
\description{
Function to take the simulated data set and format it
as required for \code{xtail}
}
