% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_update_mvn.R
\name{nc_update_mvn}
\alias{nc_update_mvn}
\title{Non-conjugate update for the parameters of
multivariate normal distribution.}
\usage{
nc_update_mvn(data, pars, Sigma, mu, differentials)
}
\arguments{
\item{data}{List of observed and known variables passed down from the
ncvi wrapper function.}

\item{pars}{List of unknown variables passed down from ncvi function.}

\item{Sigma}{Current variational covariance matrix for the MVN.}

\item{mu}{Current variational mean vector for the MVN.}

\item{differentials}{List of differentials functions, should have
'mu' and 'Sigma'.}
}
\description{
Non-conjugate update for the parameters of
multivariate normal distribution.
}
