% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{posterior.data.frame}
\alias{posterior.data.frame}
\title{#' @rdname posterior
#' @importFrom brms posterior_samples
#' @export
#'
posterior.brmsfit <- function(object, ...) {
brms::posterior_samples(object, ...)
}}
\usage{
\method{posterior}{data.frame}(
  object,
  n = 5000,
  posterior = "posterior",
  remove = c("likelihood", "loglik", "prior", "loglikelihood")
)
}
\description{
#' @rdname posterior
#' @importFrom brms posterior_samples
#' @export
#'
posterior.brmsfit <- function(object, ...) {
brms::posterior_samples(object, ...)
}
}
