% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_babel.R
\name{prep_babel}
\alias{prep_babel}
\title{Function to prepare data as required by \code{babel}}
\usage{
prep_babel(data_edgeR)
}
\value{
A list with \code{ribo}, \code{rna}, and \code{group}
}
\description{
Function to prepare data as required by \code{babel}
}
