% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_baySeq.R
\name{prep_bayseq}
\alias{prep_bayseq}
\title{Expects samples to be organized as Ribo (ctrl, trt), RNA (ctrl, trt)}
\usage{
prep_bayseq(data_edgeR)
}
\description{
Expects samples to be organized as Ribo (ctrl, trt), RNA (ctrl, trt)
}
