% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_deseq2.R
\name{prep_deseq}
\alias{prep_deseq}
\title{Function to prepare the counts table and annotations
required by \code{DESeq2}}
\usage{
prep_deseq(data_edgeR)
}
\arguments{
\item{data_edgeR}{List, as produced by \code{sim_data_mixture()}}
}
\value{
A list with \code{counts} and \code{experiment}
}
\description{
Function to prepare the counts table and annotations
required by \code{DESeq2}
}
