% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_ncvi.R
\name{prep_ncvi}
\alias{prep_ncvi}
\title{Prepare list of objects required for \code{fit_ncvi()}}
\usage{
prep_ncvi(counts, X, Z)
}
\arguments{
\item{counts}{Data frame of counts with first column the
gene ID}

\item{X}{Fixed effects design matrix}

\item{Z}{Random effects design matrix}
}
\description{
Prepare list of objects required for \code{fit_ncvi()}
}
