% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_xtail.R
\name{prep_xtail}
\alias{prep_xtail}
\title{Function to take the simulated data set and format it
as required for \code{xtail}}
\usage{
prep_xtail(data_edgeR)
}
\arguments{
\item{data_edgeR}{output element of \code{sim_data_mixture()}}
}
\value{
list with RPF and mRNA data frames and
vector of treatment conditions
}
\description{
Function to take the simulated data set and format it
as required for \code{xtail}
}
