% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data_mixture.R
\name{sim_data_mixture}
\alias{sim_data_mixture}
\title{Simulate a set of data to fit with
the REH mixture prior model using \code{fit_ncvi()} or \code{stan()}}
\usage{
sim_data_mixture(settings)
}
\description{
Simulate a set of data to fit with
the REH mixture prior model using \code{fit_ncvi()} or \code{stan()}
}
