% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_data_REH}
\alias{sim_data_REH}
\title{Simulate a set of data to fit with
the REH model using \code{fit_ncvi()} or \code{stan()}}
\usage{
sim_data_REH(settings)
}
\description{
Simulate a set of data to fit with
the REH model using \code{fit_ncvi()} or \code{stan()}
}
