% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex4_legacy.R
\name{update_EB}
\alias{update_EB}
\title{Legacy update for mvn mean}
\usage{
update_EB(data, pars, differential_EB)
}
\arguments{
\item{data}{List with \verb{etc$X, y, etc$t}}

\item{pars}{List with \verb{phi$EB, phi$SigmaB}}
}
\description{
Legacy update for mvn mean
}
