% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex4functions.R
\name{update_EM}
\alias{update_EM}
\title{Legacy update for conjugate mvn hierarchical mean}
\usage{
update_EM(data, pars)
}
\arguments{
\item{data}{List with \verb{etc$t, etc$G, etc$s}}

\item{pars}{List with \code{phi$EB}}
}
\description{
Legacy update for conjugate mvn hierarchical mean
}
