% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex4_legacy.R
\name{update_pars_legacy}
\alias{update_pars_legacy}
\title{Wrapper for legacy format updates to pass to \code{fit_ncvi()}}
\usage{
update_pars_legacy(data, pars, args)
}
\description{
Wrapper for legacy format updates to pass to \code{fit_ncvi()}
}
