% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex4_new.R
\name{update_pars_new}
\alias{update_pars_new}
\title{New wrapper update wrapper for hierarchical poisson / ex4}
\usage{
update_pars_new(data, pars, args)
}
\description{
New wrapper update wrapper for hierarchical poisson / ex4
}
