% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex4functions.R
\name{update_phi_ex4}
\alias{update_phi_ex4}
\title{Update using the legacy format/differentials for ex4}
\usage{
update_phi_ex4(data, pars, differentials)
}
\description{
Update using the legacy format/differentials for ex4
}
