% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex4_legacy.R
\name{update_phi_legacy}
\alias{update_phi_legacy}
\title{Update using the legacy format/differentials for ex4}
\usage{
update_phi_legacy(data, pars, differentials)
}
\description{
Update using the legacy format/differentials for ex4
}
