% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex4_new.R
\name{update_phi_new}
\alias{update_phi_new}
\title{New update function for phi param in hierarchical mean poisson example}
\usage{
update_phi_new(data, pars, differentials)
}
\arguments{
\item{data}{List of observed variables, with \verb{y, C, G}}

\item{pars}{List of pars for update, with \verb{phi, theta}}

\item{differentials}{List of differentials functions, with \verb{mu, Sigma}}
}
\description{
New update function for phi param in hierarchical mean poisson example
}
