% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex4functions.R
\name{update_SigmaB}
\alias{update_SigmaB}
\title{Legacy update for mvn covariance}
\usage{
update_SigmaB(data, pars, differential_SigmaB)
}
\arguments{
\item{data}{List with \verb{etc$X, y, etc$t}}

\item{pars}{List with \verb{phi$EB, phi$SigmaB}}
}
\description{
Legacy update for mvn covariance
}
