% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex4_legacy.R
\name{update_theta_legacy}
\alias{update_theta_legacy}
\title{Legacy format update for \code{theta} in ex4}
\usage{
update_theta_legacy(data, pars)
}
\description{
Legacy format update for \code{theta} in ex4
}
