% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex4_new.R
\name{update_theta_new}
\alias{update_theta_new}
\title{New update function for theta param in hierarchical example;
treats everything but \code{theta$M} as fixed}
\usage{
update_theta_new(data, pars, priors)
}
\arguments{
\item{data}{List of observed variables, with \verb{y, C, G}}

\item{pars}{List of pars for update, with \verb{phi, theta}}
}
\description{
New update function for theta param in hierarchical example;
treats everything but \code{theta$M} as fixed
}
